Slammin' for Dusbowl, Genesis, Monkey, Palermo...
And Schtop and Warpath! I did some slight updates to Schtop, so I had to build the cubemaps for it again.

NOTE that the maps which were compiled with HDR+LDR in the past are HDR-only now. Monkey may also look dark, please tell me if it looks kinda bad (like light not coming through the leaves.