It is decided, LASER grenades is the way
Elmo: "This was one of the reasons the laser/lazer gren replaced the nail gren. The explosion wasn’t skill if it got someone. It was pure luck. The reason most things changed was to make the game more skilful and intentional and/or to remove frustration/confusion for player(s) caught up in bad game mechanics"